requirements.txt
# Core Framework
streamlit==1.40.0

# LangChain & LLM
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
openai==1.58.1

# RAG - Vector Store
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
langchain-huggingface==0.1.2

# Data Processing
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5

# Document Loaders
pypdf==5.1.0
python-docx==1.1.2

# Visualization
plotly==5.24.1
matplotlib==3.9.3
seaborn==0.13.2

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
